E2E test stubs extended to journey.cy.ts
