+++ cypress/e2e/pages.cy.ts
@@ -1,69 +1,7 @@
-import { CyHttpMessages, RouteMatcher, RouteMatcherOptionsGeneric } from "cypress/types/net-stubbing";
+import { oracc_stubs } from "cypress/e2e/oracc_stubs";
 
-class MockResponse {
-  constructor(public body_fixture_path: string) {}
-}
-
-type InterceptData = {
-  response: CyHttpMessages.IncomingHttpResponse;
-  routeMatcher: RouteMatcher;
-}
-
-describe('Page', () => {
-  let searches: InterceptData[] = [];
-  let suggests: InterceptData[] = [];
-  let oracc_calls: InterceptData[] = [];
-  // These generic intercepts need to be declared before the generic
-  // ones or they won't run.
-  beforeEach(() => {
-    cy.intercept('GET', 'http://localhost:8000/search/*', (req) => {
-      req.continue((res) => {
-        searches.push(interceptData(req, res));
-      });
-    });
-    cy.intercept('GET', 'http://localhost:8000/suggest_all/*', (req) => {
-      req.continue((res) => {
-        suggests.push(interceptData(req, res));
-      });
-    });
-    cy.intercept('GET', 'https://build-oracc.museum.upenn.edu/**', (req) => {
-      req.continue((res) => {
-        oracc_calls.push(interceptData(req, res));
-      });
-    });
-  });
-  beforeEach(() => {
-    cy.fixture('directory.json').then((dir) => {
-      Object.keys(dir).forEach((fixture_path) => {
-        cy.intercept(dir[fixture_path], (req) => {
-          req.reply({ fixture: fixture_path });
-        });
-      });
-    });
-  });
-  after(() => {
-    const output_dir = 'cypress/fixtures_new';
-    const searches_dir = writeCapturedResponses(
-      searches,
-      output_dir,
-      'backend'
-    );
-    const suggests_dir = writeCapturedResponses(
-      suggests,
-      output_dir,
-      'suggests'
-    );
-    const oracc_dir = writeCapturedResponses(
-      oracc_calls,
-      output_dir,
-      'oracc'
-    );
-    cy.writeFile(`${output_dir}/directory.json`, {
-      ...searches_dir,
-      ...suggests_dir,
-      ...oracc_dir
-    });
-  });
+describe('Pages', () => {
+  oracc_stubs('pages');
 
   describe('home page', () => {
     it('is the search page', () => {
@@ -117,52 +55,3 @@ describe('Page', () => {
     });
   });
 });
-
-function interceptData(
-  req: CyHttpMessages.IncomingHttpRequest,
-  res: CyHttpMessages.IncomingHttpResponse
-): InterceptData {
-  const url = new URL(req.url);
-  return {
-    response: res,
-    routeMatcher: {
-      hostname: url.hostname,
-      pathname: url.pathname,
-      query: Object.fromEntries(url.searchParams.entries())
-    }
-  };
-}
-
-function writeCapturedResponses(
-  captured: InterceptData[],
-  fixture_dir: string,
-  name: string
-): Map<string, RouteMatcher> {
-  let r = new Map<string, RouteMatcher>();
-  captured.forEach((intercept_data) => {
-    const loc: string = urlToFilePath(
-      (intercept_data.routeMatcher as RouteMatcherOptionsGeneric<string>).pathname,
-      (intercept_data.routeMatcher as RouteMatcherOptionsGeneric<string>).query,
-      [name]
-    );
-    cy.writeFile(`${fixture_dir}/${loc}`, intercept_data.response.body);
-    r[loc] = intercept_data.routeMatcher;
-  });
-  return r;
-}
-
-function urlToFilePath(pathname: string, query: any, initial: string[]) {
-  const params = Object.keys(query).sort();
-  let qs = params.map((param) => `${param}=${query[param]}`);
-  let path = pathname.split('/').slice(1); // initial '/' produces initial ''
-  let name = 'index';
-  if (path.length) {
-    name = path.pop();
-  } else if (qs.length) {
-    name = qs.pop();
-  }
-  const steps = initial.concat(path, qs, [name]);
-  const loc = steps.join('/');
-  return loc;
-}
-
