+++ cypress/fixtures/oracc/neo/ref%3DQ003478.36/xis%3Dqpn.r00002c/qpn
@@ -446,18 +446,18 @@ GATEWAY_INTERFACE=CGI/1.1
 HTTPS=on
 HTTP_ACCEPT=application/json, text/plain, */*
 HTTP_ACCEPT_ENCODING=gzip
-HTTP_ACCEPT_LANGUAGE=en-GB,en-US;q=0.9,en;q=0.8
+HTTP_ACCEPT_LANGUAGE=en-GB
 HTTP_CONNECTION=keep-alive
 HTTP_HOST=build-oracc.museum.upenn.edu
 HTTP_ORIGIN=http://localhost:4200
 HTTP_REFERER=http://localhost:4200/
-HTTP_SEC_CH_UA="Not/A)Brand";v="8", "Chromium";v="126", "Google Chrome";v="126"
+HTTP_SEC_CH_UA="Not=A?Brand";v="99", "Chromium";v="118"
 HTTP_SEC_CH_UA_MOBILE=?0
 HTTP_SEC_CH_UA_PLATFORM="Linux"
 HTTP_SEC_FETCH_DEST=empty
 HTTP_SEC_FETCH_MODE=cors
 HTTP_SEC_FETCH_SITE=cross-site
-HTTP_USER_AGENT=Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/126.0.0.0 Safari/537.36
+HTTP_USER_AGENT=Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Cypress/13.13.1 Chrome/118.0.5993.159 Electron/27.3.10 Safari/537.36
 ORACC=/home/oracc
 ORACC_BUILDS=/home/oracc
 ORACC_HOME=/home/oracc
@@ -469,8 +469,8 @@ PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin
 PATH_INFO=/neo/qpn
 PATH_TRANSLATED=/home/oracc/www/cgi-bin/wx/neo/qpn
 QUERY_STRING=xis=qpn.r00002c&ref=Q003478.36
-REMOTE_ADDR=86.12.69.28
-REMOTE_PORT=56630
+REMOTE_ADDR=128.16.6.66
+REMOTE_PORT=47084
 REQUEST_METHOD=GET
 REQUEST_SCHEME=https
 REQUEST_URI=/neo/qpn?xis=qpn.r00002c&ref=Q003478.36
