+++ cypress/fixtures/oracc/neo/ref%3Dundefined/xis%3Dsux.r0000c5/sux
@@ -1,0 +1,234 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <link rel="stylesheet" type="text/css" href="/css/p4error.css"/>
+    <title>Oracc Pager Error</title>    
+  </head>
+  <body>
+    <h1>Oracc Pager Error</h1>
+    <p>The pager was unable to process the last request successfully.</p>
+    <p>The reported error was:</p>
+
+    <pre id="p4Error">px: error: key undefined not found in item db
+</pre>
+    
+    <div id="p4Debug">
+  <h1>Debugging Information</h1>
+
+  <h2>Invocation</h2>
+  
+  <p>The pager was invoked as:</p>
+  
+  <pre id="p4Invoked">/home/oracc/bin/px web=1 xis=sux%2Er0000c5 ref=undefined glos=sux project=neo</pre>
+
+  <h2>Pager Status</h2>
+  
+  <p>The pager reported status as:</p>
+
+  <pre id="p4Status"><h2>Internal Data Structure State (Isp *ip)</h2>
+	oracc=/home/oracc
+	from=list
+	data=dtxt
+	show=rref
+	project=neo
+	projdir=/home/oracc/neo
+	list_name=list
+	op_nlevels=0
+	dors=0
+	perm=123
+	zoom=0
+	page=1
+	psiz=25
+	srch=(null)
+	glos=sux
+	item=undefined
+	hili=(null)
+	item_replace=(null)
+	bkmk=(null)
+	ceid=xtf
+	cemd=line
+	lang=en
+	xhmd=html
+	part=(null)
+	form=(null)
+	what=pager
+	wrap=(null)
+	uimd=(null)
+	pack=asis
+	host=(null)
+	aapi=rest
+	host_path=(null)
+	sig=(null)
+	tmp_dir=(null)
+	err=px: error: key %s not found in item db
+
+	errx=undefined
+	nowhat=0
+	force=2
+	olev=0
+	debug=0
+	pub_output=0
+	verbose=0
+	web=1
+	zlev=3
+	argc=6
+	referer=(null)
+	tmpdir=(null)
+<h3>	struct isp_cache ip->cache</h3>
+		sys=/home/oracc/www/p4.d
+		project=/home/oracc/www/p4.d/neo
+		sub=/home/oracc/www/p4.d/neo/sux/sux.r0000c5
+		out=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123
+		list=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/list
+		sort=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123
+		csi=/home/oracc/pub/neo/sortinfo.csi
+		tsv=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/pag.tsv
+		max=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/max.tsv
+		mol=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/zoom.mol
+		pkey=(null)
+		pgin=(null)
+		page=(null)
+		zout=(null)
+		item=(null)
+		prox=(null)
+		meta=(null)
+		html=(null)
+		ltab=(null)
+		hilite=(null)
+		pub=/home/oracc/neo/02pub/p4.d
+		use=/home/oracc/www/p4.d/neo/sux/sux.r0000c5
+		txtindex=(null)
+		t_sort=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/t
+		t_tsv=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/t/pag.tsv
+		t_max=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/t/max.tsv
+		t_mol=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/123/123/t/zoom.mol
+<h3>	struct isp_config ip->default_cfg</h3>
+		leftmenu=1
+		select=1
+		sort_fields=period,subgenre,provenience
+		sort_labels=Time,Genre,Place
+		head_template=1 2 3
+		cat_fields=designation,primary_publication,subgenre|genre,period,place|provenience
+		cat_links=(null)
+		cat_widths=auto,17,17,17,17
+<h3>	struct isp_config ip->special_cfg</h3>
+		leftmenu=1
+		select=0
+		sort_fields=period,genre,provenience
+		sort_labels=Time,Genre,Place
+		head_template=1 2 3
+		cat_fields=designation,primary_publication,subgenre|genre,period,place|provenience
+		cat_links=(null)
+		cat_widths=auto,17,17,17,17
+<h3>	struct isp_glosdata ip->glosdata</h3>
+		dir=/home/oracc/neo/02pub/cbd/sux
+		web=(null)
+		let=entry_ids
+		lmax=(null)
+		ent=(null)
+		xis=sux.r0000c5
+		ltab=A	L000	B	L001	D	L002	E	L003	G	L004	Ŋ	L005	H	L006	I	L007	K	L008	L	L009	M	L010	N	L011	P	L012	R	L013	S	L014	Š	L015	T	L016	U	L017	Y	L018	Z	L019	
+
+		lbase=entry_ids
+		lpath=/home/oracc/pub/neo/cbd/sux/entry_ids.lst
+		ecpath=(null)
+		emax=2321
+		ipath=(null)
+<h3>	struct isp_itemdata ip->itemdata</h3>
+		langs=(null)
+		nlangs=0
+		xtflang=(null)
+		lmem=(null)
+		item=undefined
+		block=(null)
+		proj=(null)
+		htmd=(null)
+		html=(null)
+		dots=(null)
+		index=(null)
+		page=(null)
+		pindex=(null)
+		zoom=(null)
+		zpag=(null)
+		zindex=(null)
+		prev=(null)
+		next=(null)
+		tmax=1
+		xmdxsl=(null)
+		bld=(null)
+		not=0
+<h3>	struct isp_list_loc ip->lloc</h3>
+		type=xis
+		lang=sux
+		method=xisdb
+		dbpath=/home/oracc/pub/neo/cbd/sux
+		dbname=tis
+		path=/home/oracc/www/p4.d/neo/sux/sux.r0000c5/list
+<h3>	struct isp_srchdata ip->srchdata</h3>
+		tmp=(null)
+		bar=(null)
+		count=0
+		gran=(null)
+		list=(null)
+		new=0
+		adhoc=0
+	zmax=2
+</pre>
+
+  <h2>Environment</h2>
+
+  <p>HTTP environment variables:</p>
+
+  <pre id="p4Environment">CONTEXT_DOCUMENT_ROOT=/home/oracc/www
+CONTEXT_PREFIX=
+DOCUMENT_ROOT=/home/oracc/www
+GATEWAY_INTERFACE=CGI/1.1
+HTTPS=on
+HTTP_ACCEPT=application/json, text/plain, */*
+HTTP_ACCEPT_ENCODING=gzip
+HTTP_ACCEPT_LANGUAGE=en-GB
+HTTP_CONNECTION=keep-alive
+HTTP_HOST=build-oracc.museum.upenn.edu
+HTTP_ORIGIN=http://localhost:4200
+HTTP_REFERER=http://localhost:4200/
+HTTP_SEC_CH_UA="Not=A?Brand";v="99", "Chromium";v="118"
+HTTP_SEC_CH_UA_MOBILE=?0
+HTTP_SEC_CH_UA_PLATFORM="Linux"
+HTTP_SEC_FETCH_DEST=empty
+HTTP_SEC_FETCH_MODE=cors
+HTTP_SEC_FETCH_SITE=cross-site
+HTTP_USER_AGENT=Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Cypress/13.13.1 Chrome/118.0.5993.159 Electron/27.3.10 Safari/537.36
+ORACC=/home/oracc
+ORACC_BUILDS=/home/oracc
+ORACC_HOME=/home/oracc
+ORACC_HOST=build-oracc.museum.upenn.edu
+ORACC_MODE=multi
+ORACC_PROJECT=neo
+ORACC_USER=yes
+PATH=/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin:/snap/bin
+PATH_INFO=/neo/sux
+PATH_TRANSLATED=/home/oracc/www/cgi-bin/wx/neo/sux
+QUERY_STRING=xis=sux.r0000c5&ref=undefined
+REMOTE_ADDR=128.16.6.66
+REMOTE_PORT=48102
+REQUEST_METHOD=GET
+REQUEST_SCHEME=https
+REQUEST_URI=/neo/sux?xis=sux.r0000c5&ref=undefined
+SCRIPT_FILENAME=/home/oracc/www/cgi-bin/wx
+SCRIPT_NAME=
+SCRIPT_URI=https://build-oracc.museum.upenn.edu/neo/sux
+SERVER_ADDR=130.91.81.219
+SERVER_ADMIN=stinney@upenn.edu
+SERVER_NAME=build-oracc.museum.upenn.edu
+SERVER_PORT=443
+SERVER_PROTOCOL=HTTP/1.1
+SERVER_SIGNATURE=<address>Apache/2.4.52 (Ubuntu) Server at build-oracc.museum.upenn.edu Port 443</address>
+
+SERVER_SOFTWARE=Apache/2.4.52 (Ubuntu)
+SSL_TLS_SNI=build-oracc.museum.upenn.edu
+</pre>
+
+</div>
+
+  </body>
+</html>
