+++ cypress/fixtures/suggests/suggest_all/wa
@@ -1,0 +1,105 @@
+{
+  "completions": [
+    "way",
+    "wax",
+    "waṣû",
+    "warû",
+    "walk",
+    "watû",
+    "wapû",
+    "wave",
+    "wadû",
+    "wash",
+    "wail",
+    "wall",
+    "warp",
+    "wardu",
+    "waklu",
+    "warhu",
+    "warki",
+    "warkû",
+    "waqru",
+    "wašbu",
+    "water",
+    "watru",
+    "warka",
+    "wašru",
+    "wašṭu",
+    "waṣpu",
+    "wasmu",
+    "watmu",
+    "waldu",
+    "warqu",
+    "waršu",
+    "waste",
+    "waṣbu",
+    "~wall",
+    "wašābu",
+    "wabālu",
+    "waṣābu",
+    "wašāru",
+    "watāru",
+    "warādu",
+    "waqāru",
+    "walādu",
+    "wasāmu",
+    "warāqu",
+    "waters",
+    "washed",
+    "warkat",
+    "wardūtu",
+    "warkītu",
+    "warkānu",
+    "watmanu",
+    "watartu",
+    "walkway",
+    "warlike",
+    "washing",
+    "watcher",
+    "wardatu",
+    "warrior",
+    "warkatu",
+    "Walilum",
+    "wailing",
+    "watering",
+    "wasteland",
+    "waterskin",
+    "Warad-Sîn",
+    "washbasin",
+    "washerman",
+    "Warad-Sin",
+    "warhišamma",
+    "wait (for)",
+    "wash water",
+    "wash (off)",
+    "waste field",
+    "warriorhood",
+    "watercourse",
+    "water-clock",
+    "Warad-Šamaš",
+    "wall-plinth",
+    "Warad-Ištar",
+    "watch-tower",
+    "Warad-Tutub",
+    "Warad-ilišu",
+    "water-table",
+    "walk forward",
+    "was standing",
+    "waste ground",
+    "(water-)skin",
+    "ways of life",
+    "wall plaster",
+    "water drawer",
+    "(wall) hook?",
+    "Warad-Zababa",
+    "water channel",
+    "water-channel",
+    "Warad-Amurrim",
+    "walking round",
+    "wasted pieces",
+    "watering place",
+    "(wall-)plaster",
+    "(water-) meadow"
+  ],
+  "suggestions": []
+}
