+++ cypress/fixtures/suggests/suggest_all/wat
@@ -1,0 +1,36 @@
+{
+  "completions": [
+    "watû",
+    "water",
+    "watru",
+    "watmu",
+    "watāru",
+    "waters",
+    "watmanu",
+    "watartu",
+    "watcher",
+    "watering",
+    "waterskin",
+    "watercourse",
+    "water-clock",
+    "watch-tower",
+    "water-table",
+    "(water-)skin",
+    "water drawer",
+    "water channel",
+    "water-channel",
+    "watering place",
+    "(water-) meadow"
+  ],
+  "suggestions": [
+    "what",
+    "way",
+    "wet",
+    "wait",
+    "was",
+    "wax",
+    "war",
+    "wag",
+    "watu"
+  ]
+}
