+++ cypress/fixtures/suggests/suggest_all/water
@@ -1,0 +1,25 @@
+{
+  "completions": [
+    "water",
+    "waters",
+    "watering",
+    "waterskin",
+    "watercourse",
+    "water-clock",
+    "water-table",
+    "(water-)skin",
+    "water drawer",
+    "water channel",
+    "water-channel",
+    "watering place",
+    "(water-) meadow"
+  ],
+  "suggestions": [
+    "wataru",
+    "watmu",
+    "watru",
+    "watu",
+    "wa-tar",
+    "watar"
+  ]
+}
